+++ R/apa_print_glht.R
@@ -9,7 +9,7 @@
 #'    Other confidence intervals can be supplied as a \code{vector} of length 2 (lower and upper boundary, respectively)
 #'    with attribute \code{conf.level}, e.g., when calculating bootstrapped confidence intervals.
 # #' @param contrast_names Character. A vector of names to identify calculated contrasts.
-#' @param ... Further arguments to pass to \code{\link{printnum}} to format the estimate.
+#' @param ... Further arguments to pass to \code{\link{apa_num}} to format the estimate.
 #' @inheritParams glue_apa_results
 #'
 #' @evalRd apa_results_return_value()
@@ -64,7 +64,7 @@ apa_print.summary.glht <- function(
   multcomp_adjustment <- if(x$test$type == "none") multcomp::univariate_calpha() else multcomp::adjusted_calpha()
   print_ci <- stats::confint(x, level = conf.int, calpha = multcomp_adjustment)$confint
   dimnames(print_ci) <- NULL
-  table_ci <- unlist(do.call("print_confint", c(list(x = print_ci[, -1]), ellipsis))) # Remove point estimate from matrix
+  table_ci <- unlist(do.call("apa_confint", c(list(x = print_ci[, -1]), ellipsis))) # Remove point estimate from matrix
   tidy_x$std.error <- table_ci
   colnames(tidy_x)[colnames(tidy_x) == "std.error"] <- "conf.int"
 
@@ -73,12 +73,12 @@ apa_print.summary.glht <- function(
   ## Typeset columns
   sanitized_contrasts <- sanitize_terms(tidy_x$contrast)
   tidy_x$contrast <- beautify_terms(tidy_x$contrast)
-  tidy_x$estimate <- do.call("printnum", c(list(x = tidy_x$estimate), ellipsis))
-  tidy_x$statistic <- printnum(tidy_x$statistic, digits = 2)
-  tidy_x[[p_value]] <- printp(tidy_x[[p_value]])
+  tidy_x$estimate <- do.call("apa_num", c(list(x = tidy_x$estimate), ellipsis))
+  tidy_x$statistic <- apa_num(tidy_x$statistic, digits = 2)
+  tidy_x[[p_value]] <- apa_p(tidy_x[[p_value]])
 
   if(x$df != 0) {
-    tidy_x$df <- print_df(x$df)
+    tidy_x$df <- apa_df(x$df)
     tidy_x <- tidy_x[, c("contrast", "null.value", "estimate", "conf.int", "statistic", "df", p_value)] # sort columns
     variable_label(tidy_x$df) <- "$\\mathit{df}$"
   }