+++ tests/testthat/test_apa_print_htest.R
@@ -286,7 +286,7 @@ test_that(
 
     expect_identical(
       attr(two_sample_prop_test_output$table$statistic, "n")
-      , printnum(as.integer(sum(patients[3:4])))
+      , apa_num(as.integer(sum(patients[3:4])))
     )
   }
 )